@@ -52,26 +52,26 @@
     </distributionManagement>
 
     <properties>
-        <maven.compiler.target>11</maven.compiler.target>
-        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
         <jamm.version>0.4.1</jamm.version>
-        <jackson.version>2.13.2.2</jackson.version>
-        <gson.version>2.9.0</gson.version>
-        <org.json.version>20220320</org.json.version>
-        <junit.version>5.8.2</junit.version>
-        <archunit.version>0.23.1</archunit.version>
-        <sonar.version>3.7.0.1746</sonar.version>
-        <surefire.version>2.22.2</surefire.version>
-        <jacoco.version>0.8.8</jacoco.version>
-        <maven.checkstyle.version>3.1.2</maven.checkstyle.version>
+        <jackson.version>2.18.3</jackson.version>
+        <gson.version>2.12.1</gson.version>
+        <org.json.version>20250107</org.json.version>
+        <junit.version>5.13.0-M1</junit.version>
+        <archunit.version>1.4.0</archunit.version>
+        <sonar.version>5.0.0.4389</sonar.version>
+        <surefire.version>3.5.2</surefire.version>
+        <jacoco.version>0.8.12</jacoco.version>
+        <maven.checkstyle.version>3.6.0</maven.checkstyle.version>
         <maven-release-plugin.version>2.5.3</maven-release-plugin.version>
         <maven-javadoc.version>3.3.1</maven-javadoc.version>
         <maven-gpg.version>3.0.1</maven-gpg.version>
         <maven-source-plugin.version>3.2.1</maven-source-plugin.version>
-        <checkstyle.version>9.0.1</checkstyle.version>
+        <checkstyle.version>10.21.4</checkstyle.version>
         <jvm.options>
             -javaagent:${user.home}/.m2/repository/io/github/stephankoelle/jamm/${jamm.version}/jamm-${jamm.version}.jar
-            --add-opens java.base/java.util=ALL-UNNAMED --add-opens java.base/java.lang=ALL-UNNAMED --add-opens
+            --add-opens java.base/java.util=ALL-UNNAMED --add-opens jakarta.validation/jakarta.validation=ALL-UNNAMED --add-opens java.base/java.lang=ALL-UNNAMED --add-opens
             java.base/java.math=ALL-UNNAMED --add-opens json.parser/com.techatpark.sjson=ALL-UNNAMED
         </jvm.options>
     </properties>
@@ -89,6 +89,15 @@
     </dependencyManagement>
 
     <dependencies>
+
+
+        <dependency>
+            <groupId>jakarta.validation</groupId>
+            <artifactId>jakarta.validation-api</artifactId>
+            <version>3.0.2</version>
+        </dependency>
+
+
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
@@ -124,7 +133,7 @@
 
         <dependency>
             <groupId>com.tngtech.archunit</groupId>
-            <artifactId>archunit</artifactId>
+            <artifactId>archunit-junit5</artifactId>
             <version>${archunit.version}</version>
             <scope>test</scope>
         </dependency>
@@ -145,7 +154,6 @@
                     </dependency>
                 </dependencies>
                 <configuration>
-                    <encoding>UTF-8</encoding>
                     <consoleOutput>true</consoleOutput>
                     <failsOnError>true</failsOnError>
                     <linkXRef>false</linkXRef>
@@ -223,7 +231,6 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-failsafe-plugin</artifactId>
-                <version>${surefire.version}</version>
             </plugin>
 
 
