@@ -1,15 +1,19 @@
 package com.techatpark.sjson;
 
-import com.techatpark.sjson.util.NumberParser;
+import com.techatpark.sjson.element.JsonArray;
+import com.techatpark.sjson.element.JsonFalse;
+import com.techatpark.sjson.element.JsonNull;
+import com.techatpark.sjson.element.JsonNumber;
+import com.techatpark.sjson.element.JsonObject;
+import com.techatpark.sjson.element.JsonString;
+import com.techatpark.sjson.element.JsonTrue;
 
 import java.io.IOException;
 import java.io.Reader;
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+
 
 /**
  * Json parser for server side workloads.
@@ -23,36 +27,28 @@ import java.util.Map;
  * Note:
  * This is not general purpose parser. This is useful for Microservices
  * and REST Clients where we primarily need to read/write json data.
+ * @param <T> type of object
  */
-public final class Json {
+public sealed interface Json<T>
+        permits JsonString, JsonNumber, JsonTrue, JsonFalse,
+            JsonNull, JsonArray, JsonObject,
+        Json.Parser {
 
     /**
-     * Length for hex char.
+     * Length of Unicode.
      */
-    private static final int LENGTH = 4;
-
-    /**
-     * Radix for hex char.
-     */
-    private static final int RADIX = 16;
-
-    /**
-     * Capacity.
-     */
-    private static final int CAPACITY = 10;
-
-
-
-    /**
-     * Number 0.
-     */
-    private static final int NUMBER_ZERO = 0;
+    int UNICODE_LENGTH = 4;
 
     /**
      * For invalid JSON.
      */
-    private static final String ILLEGAL_JSON_VALUE = "Illegal value at ";
+    String ILLEGAL_JSON_VALUE = "Illegal value at ";
 
+    /**
+     * Reads the object value.
+     * @return value
+     */
+    T read();
 
     /**
      * Reads JSON as a Java Object.
@@ -72,9 +68,19 @@ public final class Json {
      * @return object
      * @throws IOException - throws io exception
      */
-    public Object read(final Reader reader) throws IOException {
+    static Object read(final Reader reader) throws IOException {
+        return parse(reader).read();
+    }
+
+    /**
+     * Parses the reader for Json.
+     * @param reader
+     * @return json
+     * @throws IOException
+     */
+    private static Json<?> parse(final Reader reader) throws IOException {
         try (reader) {
-            return new ContentExtractor(reader).getValue();
+            return new Parser(reader).parse();
         }
     }
 
@@ -84,26 +90,12 @@ public final class Json {
      * @param jsonMap
      * @return jsonText
      */
-    public String jsonText(final Map<String, Object> jsonMap) {
-        StringBuilder builder = new StringBuilder();
-        boolean isFirst = true;
-        builder.append("{");
-        for (Map.Entry<String, Object> entry : jsonMap.entrySet()) {
-            if (isFirst) {
-                isFirst = false;
-            } else {
-                builder.append(",");
-            }
-            // Create Key enclosed with "
-            builder.append("\"")
-                    .append(escapeJsonTxt(entry.getKey()))
-                    .append("\":"); // Create Key value separator
-
-            Object value = entry.getValue();
-
-            valueText(builder, value);
-        }
-        return builder.append("}").toString();
+    static String jsonText(final Map<String, Object> jsonMap) {
+        return "{" + jsonMap.entrySet().stream()
+                .map(entry ->
+                        "\"" + escapeJsonTxt(entry.getKey()) + "\":"
+                        + getValue(entry.getValue()
+                        )).collect(Collectors.joining(",")) + "}";
     }
 
     /**
@@ -112,55 +104,26 @@ public final class Json {
      * @param jsonArray
      * @return jsonTxt
      */
-    public String jsonText(final List<Object> jsonArray) {
-        final StringBuilder builder = new StringBuilder("[");
-        boolean isFirst = true;
-        for (Object value: jsonArray) {
-            if (isFirst) {
-                isFirst = false;
-            } else {
-                builder.append(",");
-            }
-            valueText(builder, value);
-        }
-        return builder.append("]").toString();
+    private static String jsonText(final List<Object> jsonArray) {
+        return "[" + jsonArray.stream()
+                .map(Json::getValue)
+                .collect(Collectors.joining(",")) + "]";
     }
 
     /**
      * Create Value in according to the Type.
      *
-     * @param builder
      * @param value
+     * @return valueText
      */
-    private void valueText(final StringBuilder builder, final Object value) {
-        if (value == null) {
-            builder.append("null");
-        } else if (value instanceof String) {
-            processString(builder, (String) value);
-        } else if (value instanceof Map) {
-            builder.append(jsonText((Map<String, Object>)
-                    value));
-        } else if (value instanceof List) {
-            builder.append(jsonText((List)
-                    value));
-        } else {
-            builder.append(value);
-        }
-    }
-
-
-
-    /**
-     * Process String.
-     *
-     * @param builder
-     * @param value
-     */
-    private void processString(final StringBuilder builder,
-                               final String value) {
-        builder.append("\"")
-                .append(escapeJsonTxt(value))
-                .append("\"");
+    private static String getValue(final Object value) {
+        return switch (value) {
+            case null -> "null";
+            case String str -> "\"" + escapeJsonTxt(str) + "\"";
+            case Map map -> jsonText(map);
+            case List list -> jsonText(list);
+            default -> value.toString();
+        };
     }
 
     /**
@@ -170,7 +133,7 @@ public final class Json {
      * @param s
      * @return escapeJsonTxt
      */
-    private String escapeJsonTxt(final String s) {
+    private static String escapeJsonTxt(final String s) {
         if (s == null) {
             return null;
         }
@@ -184,7 +147,7 @@ public final class Json {
      * @param s - Must not be null.
      * @param sb
      */
-    private void escape(final String s, final StringBuilder sb) {
+    private static void escape(final String s, final StringBuilder sb) {
         final int len = s.length();
         for (int i = 0; i < len; i++) {
             char ch = s.charAt(i);
@@ -220,7 +183,7 @@ public final class Json {
                             || (ch >= '\u2000' && ch <= '\u20FF')) {
                         String ss = Integer.toHexString(ch);
                         sb.append("\\u");
-                        for (int k = 0; k < NumberParser.NUMBER_FOUR
+                        for (int k = 0; k < UNICODE_LENGTH
                                 - ss.length(); k++) {
                             sb.append('0');
                         }
@@ -233,10 +196,15 @@ public final class Json {
     }
 
     /**
-     * ContentExtractor is responsible to interact with underlying reader to
+     * Parser is responsible to interact with underlying reader to
      * extract the content.
      */
-    private final class ContentExtractor {
+    final class Parser implements Json<Object> {
+
+        /**
+         * Max Depth of an nested Object.
+         */
+        public static final int DEFAULT_MAX_DEPTH = 1000;
 
         /**
          * Reader to the JSON Content.
@@ -249,12 +217,56 @@ public final class Json {
         private char cursor;
 
         /**
+         * Depth of the current Object.
+         */
+        private short objectDepth;
+
+        /**
          * Creates Content extracter for the reader.
          *
          * @param theReader
          */
-        private ContentExtractor(final Reader theReader) {
+        Parser(final Reader theReader) {
             this.reader = theReader;
+            this.objectDepth = 0;
+        }
+
+        /**
+         * Called when new Object starts.
+         */
+        public void startObject() {
+            checkMaxDepth();
+        }
+
+        /**
+         * Called when new Object completed.
+         */
+        public void endObject() {
+            decrementDepth();
+        }
+
+        /**
+         * Called when new Array starts.
+         */
+        public void startArray() {
+            checkMaxDepth();
+        }
+        /**
+         * Called when new Array completed.
+         */
+        public void endArray() {
+            decrementDepth();
+        }
+
+        private void checkMaxDepth() {
+            if (++objectDepth == DEFAULT_MAX_DEPTH) {
+                throw new IllegalArgumentException("Document nesting depth ("
+                        + DEFAULT_MAX_DEPTH + ") exceeds the maximum allowed");
+            }
+        }
+
+        private void decrementDepth() {
+            this.objectDepth--;
         }
 
         /**
@@ -265,108 +277,85 @@ public final class Json {
          * @return object
          * @throws IOException
          */
-        private Object getValue() throws IOException {
+        public Json<?> parse() throws IOException {
             // 1. move to the first clean character to determine the Data type
             final char character = nextClean();
+            setCursor(character);
             // 2. Call corresponding get methods based on the type
-            switch (character) {
-                case '"':
-                    return getString();
-                case 'n':
-                    return getNull();
-                case 't':
-                    return getTrue();
-                case 'f':
-                    return getFalse();
-                case '{':
-                    return getObject();
-                case '[':
-                    return getArray();
-                case ']':
-                    return this;
-                default:
-                    return getNumber(character);
-            }
+            return switch (character) {
+                case '"' -> new JsonString(this);
+                case 'n' -> new JsonNull(this);
+                case 't' -> new JsonTrue(this);
+                case 'f' -> new JsonFalse(this);
+                case '{' -> new JsonObject(this);
+                case '[' -> new JsonArray(this);
+                case ']' -> this;
+                default -> {
+                    if (Character.isDigit(character)
+                            || character == '+' || character == '-') {
+                        yield new JsonNumber(this, reader, character);
+                    }
+                    throw new IllegalArgumentException(ILLEGAL_JSON_VALUE);
+                }
+            };
         }
 
         /**
-         * Reads String from Reader. Reader will stop at the " symbol
-         *
-         * @return string
+         * Sets Cursor to next clean Character.
+         */
+        public void setCursorToNextClean() throws IOException {
+            setCursor(nextClean());
+        }
+
+        /**
+         * Sets Cursor at given Character.
+         * @param character
+         */
+        public void setCursor(final char character) {
+            cursor = character;
+        }
+
+        /**
+         * Gets Cursor.
+         * @return cursor
+         */
+        public char getCursor() {
+            return cursor;
+        }
+
+        /**
+         * Skip Spaces and land reader at the valid character.
+         * @return valid character
          * @throws IOException
          */
-        private String getString() throws IOException {
-            final StringBuilder sb = new StringBuilder();
+        public char nextClean() throws IOException {
             char character;
+            do {
+                character = (char) reader.read();
+            } while (isSpace(character));
+            return character;
+        }
 
-            while ((character = getCharacter(reader.read())) != '\\'
-                    && character != '"') {
-                sb.append(character);
-            }
-
-            // Normal String
-            if (character == '"') {
-                return sb.toString();
-            }
-
-            // String with escape characters ?!
-            for (;;) {
-                switch (character) {
-                    case 0:
-                    case '\n':
-                    case '\r':
-                        throw new IllegalArgumentException(ILLEGAL_JSON_VALUE);
-                    case '\\':
-                        character = getCharacter(reader.read());
-                        switch (character) {
-                            case 'b':
-                                sb.append('\b');
-                                break;
-                            case 't':
-                                sb.append('\t');
-                                break;
-                            case 'n':
-                                sb.append('\n');
-                                break;
-                            case 'f':
-                                sb.append('\f');
-                                break;
-                            case 'r':
-                                sb.append('\r');
-                                break;
-                            case 'u':
-                                sb.append((char) Integer
-                                        .parseInt(new String(next(LENGTH)),
-                                                RADIX));
-                                break;
-                            case '"':
-                            case '\'':
-                            case '\\':
-                            case '/':
-                                sb.append(character);
-                                break;
-                            default:
-                                throw new IllegalArgumentException(
-                                        ILLEGAL_JSON_VALUE);
-                        }
-                        break;
-                    default:
-                        if (character == '"') {
-                            return sb.toString();
-                        }
-                        sb.append(character);
-                }
-                character = getCharacter(reader.read());
-            }
+        /**
+         * Reads next chars for given length
+         * from the reader and fill an char array.
+         * @param length
+         * @return char array
+         * @throws IOException
+         */
+        public char[] next(final int length) throws IOException {
+            char[] cbuf = new char[length];
+            reader.read(cbuf, 0, length);
+            return cbuf;
         }
 
         /**
          * get Character.
          * @throws IllegalArgumentException if EOF
-         * @param value
          * @return char value
          */
-        private char getCharacter(final int value) {
+        public char getCharacter() throws IOException {
+            final int value = reader.read();
             if (value == -1) {
                 throw new IllegalArgumentException(ILLEGAL_JSON_VALUE);
             }
@@ -374,260 +363,16 @@ public final class Json {
         }
 
         /**
-         * Reads the number from reader.
-         * Reader will stop at the next to the end of number.
-         *
-         * @param startingChar
-         * @return number
-         * @throws IOException
-         */
-        private Number getNumber(final char startingChar) throws IOException {
-
-            final StringBuilder builder = new StringBuilder(10);
-            char character;
-
-            // Happy Case : Read AllDigits before . character
-            while ((character = (char) reader.read()) != ','
-                    && Character.isDigit(character)
-                    && character != '.'
-                    && character != '}'
-                    && character != ']'
-                    && character != 'e'
-                    && character != 'E'
-                    && !isSpace(character)) {
-                builder.append(character);
-            }
-
-            // Maybe a double ?!
-            if (character == '.' || character == 'e' || character == 'E') {
-                // Decimal Number
-                if (character == '.') {
-                    StringBuilder decimals = new StringBuilder(CAPACITY);
-                    while ((character = (char) reader.read()) != ','
-                            && (Character.isDigit(character)
-                            || character == '-'
-                            || character == 'e'
-                            || character == 'E')
-                            && character != '}'
-                            && character != ']'
-
-                            && !isSpace(character)) {
-                        decimals.append(character);
-                    }
-                    setCursor(character);
-                    return getDecimalNumber(startingChar, builder, decimals);
-                } else { // Exponential Non Decimal Number
-                    builder.append(character);
-                    while ((character = (char) reader.read()) != ','
-                            && (Character.isDigit(character)
-                            || character == '-'
-                            || character == 'e'
-                            || character == 'E')
-                            && character != '}'
-                            && character != ']'
-                            && !isSpace(character)) {
-                        builder.append(character);
-                    }
-                    setCursor(character);
-                    return getExponentialNumber(startingChar, builder);
-                }
-            } else {
-                setCursor(character);
-                return getNumber(startingChar, builder);
-            }
-        }
-
-        /**
-         * Gets Number from the String.
-         *
-         * @param startingChar
-         * @param builder
-         * @return number
-         */
-        private Number getNumber(final char startingChar,
-                                 final StringBuilder builder) {
-            switch (startingChar) {
-                case '-':
-                    return NumberParser.parseNumber(builder.toString(), true);
-                case '+':
-                    return NumberParser.parseNumber(builder.toString(), false);
-                default:
-                    return NumberParser.parseNumber(builder
-                            .insert(0, startingChar)
-                            .toString(), false);
-            }
-        }
-
-        /**
-         * Gets Decimal Exponential from the String.
-         *
-         * @param startingChar
-         * @param builder
-         * @return number
-         */
-        private Number getExponentialNumber(final char startingChar,
-                                            final StringBuilder builder) {
-            return new BigDecimal(startingChar + builder.toString());
-        }
-
-        /**
-         * Gets Decimal Number from the String.
-         *
-         * @param decimal
-         * @param startingChar
-         * @param builder
-         * @return number
-         */
-        private Number getDecimalNumber(final char startingChar,
-                                        final StringBuilder builder,
-                                        final StringBuilder decimal) {
-            return NumberParser.parseDecimalNumber(startingChar
-                    + builder.toString() + "." + decimal.toString());
-        }
-
-        /**
-         * Reads True from Reader. Reader will stip at the "e" symbol.
-         *
-         * @return string
-         * @throws IOException
-         */
-        private boolean getTrue() throws IOException {
-            char[] charBuffer = next(NumberParser.NUMBER_THREE);
-            if (charBuffer[0] == 'r'
-                    && charBuffer[1] == 'u'
-                    && charBuffer[2] == 'e') {
-                // cursor = 'e';
-                return true;
-            } else {
-                throw new IllegalArgumentException(ILLEGAL_JSON_VALUE);
-            }
-        }
-
-        /**
-         * Reads False from Reader. Reader will strip at the "e" symbol.
-         *
-         * @return string
-         * @throws IOException
-         */
-        private boolean getFalse() throws IOException {
-            char[] charBuffer = next(NumberParser.NUMBER_FOUR);
-            if (charBuffer[NUMBER_ZERO] == 'a'
-                    && charBuffer[NumberParser.NUMBER_ONE] == 'l'
-                    && charBuffer[NumberParser.NUMBER_TWO] == 's'
-                    && charBuffer[NumberParser.NUMBER_THREE] == 'e') {
-                // cursor = 'e';
-                return false;
-            } else {
-                throw new IllegalArgumentException(ILLEGAL_JSON_VALUE);
-            }
-        }
-
-        /**
-         * Reads Null from Reader. Reader will stip at the "l" symbol.
-         *
-         * @return string
-         * @throws IOException
-         */
-        private Object getNull() throws IOException {
-            char[] charBuffer = next(NumberParser.NUMBER_THREE);
-            if (charBuffer[NUMBER_ZERO] == 'u'
-                    && charBuffer[NumberParser.NUMBER_ONE] == 'l'
-                    && charBuffer[NumberParser.NUMBER_TWO] == 'l') {
-                // cursor = 'l';
-                return null;
-            } else {
-                throw new IllegalArgumentException(ILLEGAL_JSON_VALUE);
-            }
-        }
-
-        /**
-         * Reads next chars for given length
-         * from the reader and fill an char array.
-         *
-         * @param length
-         * @return char array
-         * @throws IOException
-         */
-        private char[] next(final int length) throws IOException {
-            char[] cbuf = new char[length];
-            reader.read(cbuf, NUMBER_ZERO, length);
-            return cbuf;
-        }
-
-        /**
-         * Reads Object from a reader. Reader will
-         * stop at the next clean char after object.
-         *
-         * @return json object
-         * @throws IOException
-         */
-        private Map<String, Object> getObject() throws IOException {
-
-            boolean eoo = endOfObject();
-            // This is Empty Object
-            if (eoo) {
-                nextClean();
-                return Collections.emptyMap();
-            }
-            final Map<String, Object> jsonMap = new HashMap<>();
-            while (!eoo) {
-                jsonMap.put(getKey(), getValue());
-                eoo = endOfObject();
-            }
-            nextClean();
-            return Collections.unmodifiableMap(jsonMap);
-        }
-
-        /**
-         * Read Key as a String. It gets key from String Pool.
-         *
-         * @return key
-         * @throws IOException
-         */
-        private String getKey() throws IOException {
-            final String key = getString().intern();
-            nextClean();
-            return key;
-        }
-
-        /**
-         * Reades an Array. Reader stops at next clean character.
-         *
-         * @return list
-         * @throws IOException
-         */
-        private List<Object> getArray() throws IOException {
-            final Object value = getValue();
-            // If not Empty Array
-            if (value == this) {
-                nextClean();
-                return Collections.emptyList();
-            }
-            final List<Object> list = new ArrayList<>();
-            list.add(value);
-            boolean eoa = endOfArray();
-            while (!eoa) {
-                list.add(getValue());
-                eoa = endOfArray();
-            }
-            nextClean();
-            return Collections.unmodifiableList(list);
-        }
-
-
-        /**
          * Skip Spaces and land reader at the valid character.
-         *
-         * @return valid character
+         * @param character
          * @throws IOException
          */
-        private char nextClean() throws IOException {
-            char character;
+        public void next(
+                                final char character) throws IOException {
+            int charVal;
             do {
-                character = (char) this.reader.read();
-            } while (isSpace(character));
-            setCursor(character);
-            return cursor;
+                charVal = (char) reader.read();
+            } while (charVal != character);
         }
 
         /**
@@ -636,66 +381,17 @@ public final class Json {
          * @param character
          * @return flag
          */
-        private boolean isSpace(final char character) {
+        public boolean isSpace(final char character) {
             return (character == ' '
                     || character == '\n'
                     || character == '\r'
                     || character == '\t');
         }
 
-        /**
-         * Determines the Object End. By moving till " or }.
-         *
-         * @return flag
-         * @throws IOException
-         */
-        private boolean endOfObject() throws IOException {
-            char character;
-            if (cursor == '}') {
-                return true;
-            }
-            if (cursor == ',') {
-                while (getCharacter(this.reader.read()) != '"') {
-                    continue;
-                }
-                return false;
-            }
-            while ((character = getCharacter(this.reader.read())) != '"'
-                    && character != '}') {
-                continue;
-            }
-            return character == '}';
+        @Override
+        public Object read() {
+            throw new UnsupportedOperationException("Can not read Object "
+                    + "from context extractor");
         }
-
-        /**
-         * Determine array close character.
-         *
-         * @return flag
-         * @throws IOException
-         */
-        private boolean endOfArray() throws IOException {
-            char character;
-            if (cursor == ']') {
-                return true;
-            }
-            if (cursor == ',') {
-                return false;
-            }
-            while ((character = getCharacter(this.reader.read())) != ','
-                    && character != ']') {
-                continue;
-            }
-            return character == ']';
-        }
-
-        /**
-         * Sets Cursor at given Character.
-         * @param character
-         */
-        private void setCursor(final Character character) {
-            cursor = character;
-        }
-
-
     }
 }
