@@ -1,4 +1,4 @@
-package com.techatpark.sjson;
+package com.techatpark.sjson.archunit;
 
 import com.tngtech.archunit.core.domain.JavaClasses;
 import com.tngtech.archunit.core.importer.ClassFileImporter;
@@ -13,10 +13,10 @@ import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.classes;
 /**
  * This will make sure we have all the review comments met.
  */
-public class CodeReviewTest {
+class CodeReviewTest {
 
     @Test
-    public void some_architecture_rule() {
+    void some_architecture_rule() {
         JavaClasses classes = new ClassFileImporter()
                 .withImportOption(ImportOption.Predefined.DO_NOT_INCLUDE_TESTS)
                 .withImportOption(ImportOption.Predefined.DO_NOT_INCLUDE_JARS)
@@ -26,10 +26,14 @@ public class CodeReviewTest {
 
         ArchRule rule = classes().that().resideInAPackage("com.techatpark.sjson")
                 .should().onlyDependOnClassesThat()
-                .resideInAnyPackage("java.io","java.math"
+                .resideInAnyPackage("java.io"
+                        ,"java.math"
                         ,"java.util"
-                        ,"java.lang",""
-                ,"com.techatpark.sjson",
+                        , "java.util.stream"
+                        ,"jakarta.validation"
+                        ,"java.lang"
+                        ,"java.lang.reflect"
+                        ,"com.techatpark.sjson",
                         "com.techatpark.sjson.*");
 
         rule.check(classes);
